+++ src/modules/clock/main.c
@@ -42,7 +42,7 @@ void clock_unregisterinterface(struct clock_interface *interface)
 
 }
 
-void clock_initinterface(struct clock_interface *interface, unsigned char (*getseconds)(void), unsigned char (*getminutes)(void), unsigned char (*gethours)(void), unsigned char (*getweekday)(void), unsigned char (*getday)(void), unsigned char (*getmonth)(void), unsigned short (*getyear)(void))
+void clock_initinterface(struct clock_interface *interface, unsigned char (*getseconds)(void), unsigned char (*getminutes)(void), unsigned char (*gethours)(void), unsigned char (*getweekday)(void), unsigned char (*getday)(void), unsigned char (*getmonth)(void), unsigned int (*getyear)(void))
 {
 
     resource_init(&interface->resource, RESOURCE_CLOCKINTERFACE, interface);
