+++ src/modules/ethernet/ethernet.h
@@ -39,7 +39,7 @@ struct ethernet_protocol
     struct system_node root;
     struct system_node data;
     struct list datalinks;
-    unsigned short type;
+    unsigned int type;
     void (*notify)(struct ethernet_interface *interface, struct ethernet_header *header, void *buffer, unsigned int count);
 
 };
@@ -51,4 +51,4 @@ void ethernet_registerprotocol(struct ethernet_protocol *protocol);
 void ethernet_unregisterinterface(struct ethernet_interface *interface);
 void ethernet_unregisterprotocol(struct ethernet_protocol *protocol);
 void ethernet_initinterface(struct ethernet_interface *interface, unsigned int (*send)(void *buffer, unsigned int count));
-void ethernet_initprotocol(struct ethernet_protocol *protocol, char *name, unsigned short type, void (*notify)(struct ethernet_interface *interface, struct ethernet_header *header, void *buffer, unsigned int count));
+void ethernet_initprotocol(struct ethernet_protocol *protocol, char *name, unsigned int type, void (*notify)(struct ethernet_interface *interface, struct ethernet_header *header, void *buffer, unsigned int count));
